chip layouts and centering
